requirements.txt
asgiref==3.7.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.8
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-filter==23.4
django-import-export==3.3.3
django-phonenumber-field==7.2.0
django-tables2==2.6.0
django-tinymce==3.6.1
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
pdf2image==1.16.3
phonenumbers==8.13.25
Pillow==10.1.0
poppler-utils==0.1.0
python-slugify==8.0.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.8.0
xlrd==2.0.1
xlwt==1.3.0
